refactor: 全量去Linux化 + 进程调度健壮性修复 + HvFS持久化
去Linux化:
- VFS: VfsDirent→VfsDirEntry, inode_num→node_id, .inode→.node
- RamFS: RamFsInode→RamFsNode, RamFsDirent→RamFsDirEntry
  inode_bitmap→node_bitmap, root_inode→root_node, free_inodes→free_nodes
  RAMFS_MAX_INODES→RAMFS_MAX_NODES, alloc_inode→alloc_node
  内部100+处inode→node
- procfs: "self"→"current", "cpuinfo"→"sys/cpu", "meminfo"→"sys/memory"
- 用户态: UserDirent→UserDirEntry, .inode→.node
- Slab: 移除Linux对比注释
- FFI层: 所有inode_num参数名→node_id, total_inodes→total_nodes

进程/调度修复 (10项):
- 双调度器竞态消除 (SCHEDULER_EX仅做tick_accounting)
- Thread/ThreadNode UB修复 (合并为统一Thread结构)
- 中断安全 (schedule入口cli/sti)
- 内存泄漏修复 (PCB Box::from_raw释放)
- 阻塞sleep (替代忙等timer_sleep_busy)
- fork()完整实现 + clone_user_page_table
- FdTable增强 (alloc_fd/get_global_fd/close_fd)
- process_kill/孤儿回收/ryg_sysreq权限修复

HvFS持久化:
- write→sync→uberblock→reboot→mount→restore完整链路
- serialize/deserialize_dataset_metadata
- ARC key collision修复
- host-tests持久化往返测试

构建: 0 errors | 测试: 69/69 全部通过

+++ src/kernel/tests/arch.rs
@@ -81,7 +81,7 @@ fn tss_ist_operations() -> TestResult {
 fn tss_iomap() -> TestResult {
     let mut tss = TaskStateSegment::zeroed();
     check!(!tss.has_iomap(), "should not have iomap by default");
-    let offset = (TSS_SIZE / 2) as u32;
+    let offset = (TSS_SIZE / 2) as u16;
     tss.enable_iomap(offset);
     check!(tss.has_iomap(), "should have iomap after enable");
     assert_eq_test!(tss.iomap_base, offset, "iomap_base mismatch");
